+++ src/main/java/com/TH/calculator/interfaces/BasicOperations.java
@@ -1,9 +1,9 @@
 package com.TH.calculator.interfaces;
 
 public interface BasicOperations {
-	public double addition(double firstNumber, double secondNumber);
-	public double subtraction(double firstNumber, double secondNumber);
-	public double multiplication(double firstNumber, double secondNumber);
+	public int addition(int firstNumber, int secondNumber);
+	public int subtraction(int firstNumber, int secondNumber);
+	public int multiplication(int firstNumber, int secondNumber);
 	public double division(double firstNumber, double secondNumber);
 
 }
